reorganize and implement the `vertex_indices` function
