+++ rust/healpix-geo-core/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod ellipsoid;
 pub mod geometry;
 pub mod index;
-mod mesh;
 pub mod scalar;
 pub mod vectorized;
