+++ rust/healpix-geo-core/src/scalar/mod.rs
@@ -3,3 +3,4 @@ pub mod ring;
 pub mod zuniq;
 
 pub mod geometry;
+pub mod mesh;
