+++ rust/healpix-geo-core/src/scalar/nested/mod.rs
@@ -2,3 +2,4 @@ pub mod conversion;
 pub mod coordinates;
 pub mod coverage;
 pub mod hierarchy;
+pub mod mesh;
